split custom storage file
